Update npm packages and move from brunch to esbuild
