+++ apps/pathfinder_web/assets/js/board/view.jsx
