+++ apps/pathfinder_web/assets/js/controls.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import CopyToClipboard from 'react-copy-to-clipboard';
+import CopyToClipboard from 'react-copy-to-clipboard-ts';
 import {
   PLACE_WALL,
   PLACE_GOAL
