+++ apps/pathfinder_web/assets/js/text.jsx
