+++ apps/pathfinder_web/assets/package.json
@@ -1,28 +1,13 @@
 {
-  "repository": {},
-  "license": "MIT",
-  "scripts": {
-    "deploy": "brunch build --production",
-    "watch": "brunch watch --stdin"
-  },
   "dependencies": {
-    "bootstrap": "^3.3.7",
-    "jquery": "^3.2.1",
-    "mobx": "^3.2.2",
-    "mobx-react": "^4.2.2",
+    "bootstrap": "^5.3.7",
+    "jquery": "^3.7.1",
+    "mobx": "^6.13.7",
+    "mobx-react": "^9.2.0",
     "phoenix": "file:../../../deps/phoenix",
     "phoenix_html": "file:../../../deps/phoenix_html",
-    "react": "^15.6.1",
-    "react-copy-to-clipboard": "^5.0.0",
-    "react-dom": "^15.6.1"
-  },
-  "devDependencies": {
-    "babel-brunch": "6.0.6",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "brunch": "2.10.7",
-    "clean-css-brunch": "2.10.0",
-    "uglify-js-brunch": "2.1.1"
+    "react": "^19.1.0",
+    "react-copy-to-clipboard-ts": "^1.3.0",
+    "react-dom": "^19.1.0"
   }
 }
