+++ apps/pathfinder_web/config/config.exs
@@ -22,6 +22,15 @@ config :logger, :console,
   format: "$time $metadata[$level] $message\n",
   metadata: [:request_id]
 
+# Configure esbuild (the version is required)
+config :esbuild,
+  version: "0.25.5",
+  default: [
+    args: ~w(js/app.jsx --bundle --target=es2016 --outdir=../priv/static/assets),
+    cd: Path.expand("../assets", __DIR__),
+    env: %{"NODE_PATH" => Path.expand("../deps", __DIR__)}
+  ]
+
 config :phoenix, :json_library, Jason
 
 # Import environment specific config. This must remain at the bottom
