+++ apps/pathfinder_web/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "react-copy-to-clipboard-ts": "^1.3.0"
+  }
+}
